fixed spleefconfig and damage
